requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
asgiref==3.3.4
    # via django
beautifulsoup4==4.9.3
    # via UCF-Search-Service-Django (setup.py)
boto3==1.17.85
    # via UCF-Search-Service-Django (setup.py)
botocore==1.20.85
    # via
    #   boto3
    #   s3transfer
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via requests
django==3.2.13
    # via
    #   UCF-Search-Service-Django (setup.py)
    #   django-cors-headers
    #   django-filter
    #   django-mysql
    #   djangorestframework
django-cors-headers==3.7.0
    # via UCF-Search-Service-Django (setup.py)
django-filter==2.4.0
    # via UCF-Search-Service-Django (setup.py)
django-mysql==3.11.1
    # via UCF-Search-Service-Django (setup.py)
django-widget-tweaks==1.4.8
    # via UCF-Search-Service-Django (setup.py)
djangorestframework==3.12.4
    # via UCF-Search-Service-Django (setup.py)
drf-dynamic-fields==0.3.1
    # via UCF-Search-Service-Django (setup.py)
fuzzywuzzy==0.18.0
    # via UCF-Search-Service-Django (setup.py)
gunicorn==20.1.0
    # via UCF-Search-Service-Django (setup.py)
idna==2.10
    # via requests
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
lxml==4.6.5
    # via UCF-Search-Service-Django (setup.py)
mysqlclient==2.0.3
    # via UCF-Search-Service-Django (setup.py)
pillow==9.0.1
    # via UCF-Search-Service-Django (setup.py)
progress==1.5
    # via UCF-Search-Service-Django (setup.py)
# psycopg2==2.9.3
    # via UCF-Search-Service-Django (setup.py)
python-dateutil==2.8.1
    # via
    #   UCF-Search-Service-Django (setup.py)
    #   botocore
python-levenshtein==0.12.2
    # via UCF-Search-Service-Django (setup.py)
pytz==2021.1
    # via django
raven==6.10.0
    # via UCF-Search-Service-Django (setup.py)
requests==2.25.1
    # via UCF-Search-Service-Django (setup.py)
s3transfer==0.4.2
    # via boto3
six==1.16.0
    # via python-dateutil
soupsieve==2.2.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via
    #   UCF-Search-Service-Django (setup.py)
    #   django
tabulate==0.8.9
    # via UCF-Search-Service-Django (setup.py)
unidecode==1.2.0
    # via UCF-Search-Service-Django (setup.py)
urllib3==1.26.5
    # via
    #   botocore
    #   requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
